disha-backend/requirements.txt
fastapi
pydantic
firebase-admin
google-generativeai
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
uvicorn
pip-audit